+++ src/LamaPon/Graphics/GraphicsResource.h
@@ -56,7 +56,11 @@ namespace LamaPon
         Rgba16Unorm,
         Rgba16Snorm,
         R8g8B8g8Unorm,
-        G8r8G8b8Unorm
+        G8r8G8b8Unorm,
+        R8Snorm,
+        R16Snorm,
+        R11g11b10Float,
+        R9g9b9e5SharedExp
     };
 
     // Immutableは全mipの初期dataを生成時に渡す静的texture、PerMipUpdateは